Extract siloctl to its own crate; gate datafusion behind `server` feature
siloctl is a thin gRPC admin client and never needed datafusion, but it
lived in the silo crate which depended on datafusion unconditionally,
making `cargo build --bin siloctl` slow.

Two coordinated changes:

* Make `datafusion` and `datafusion-sql` optional, gated by a new
  `server` feature (in default features). Modules that use datafusion
  (`arrow_ipc`, `query`, `cluster_query`, `webui`, `server`) and the
  silo daemon binary are gated. `query_engine`-related code in
  `JobStoreShard` and `ClusterClient` is gated. Pure-Rust pb conversion
  helpers and metadata-key constants that previously lived in
  `server.rs` are extracted to `pb_convert.rs` so non-server callers
  (cluster_client, routing_client) can still reach them.
* Move `siloctl` to a dedicated workspace crate that depends on `silo`
  with `default-features = false`, so its build skips datafusion
  entirely. The existing siloctl integration tests stay in
  `silo/tests/` (they need the server) and import from the new crate
  via a dev-dependency.

`cargo build -p siloctl` no longer pulls datafusion into its dep tree;
`silo-compactor` (which already used `default-features = false`) also
benefits from the gate. Default `cargo build` of the silo daemon is
unchanged.
