+++ siloctl/Cargo.toml
@@ -1,0 +1,27 @@
+[package]
+name = "siloctl"
+version = "0.2.0"
+edition = "2024"
+description = "Administration CLI for Silo job queue clusters"
+
+[lib]
+path = "src/lib.rs"
+
+[[bin]]
+name = "siloctl"
+path = "src/main.rs"
+
+[dependencies]
+# silo without `server` feature avoids the datafusion compile cost; siloctl
+# only needs cluster_client::AuthInterceptor, pb types, settings, shard_range.
+silo = { path = "..", default-features = false }
+
+clap = { version = "4.3", features = ["derive", "env"] }
+tonic = { version = "0.12", features = ["transport"] }
+tokio = { version = "1.47.1", features = ["macros", "rt-multi-thread"] }
+serde_json = "1.0"
+rmp-serde = "1.3"
+flate2 = "1"
+anyhow = "1"
+hex = "0.4"
+base64 = "0.22"
